⚡️ Add time feature to RestValidationsBuilder
✅ Use this to check against the provided response time value

+++ src/main/java/com/shaft/validation/NativeValidationsBuilder.java
@@ -21,7 +21,7 @@ public class NativeValidationsBuilder {
 
     protected Object response;
     protected String jsonPath;
-    protected double responseTime;
+    protected long responseTime;
 
     protected String folderRelativePath;
     protected String fileName;
